Split char, string, and raw string literal CsonSerializerTests to their own dedicated test class, reducing the scope of CsonSerializerTests.
